Fix problems with consecutive symbols and static
The Lexer::gettok() function was not saving its state correctly which led to the
incorrect parsing of consecutive symbols. This commit adds a new private
variable to the lexer to save its state. Additionally, the main parser object
which is created in the main loop must be passed by reference to avoid copying
it which leads to loosing state when returning to the caller function.