Add initial JIT support
This does not compile on my machine because I did not build LLVM with rtti
support. To do this, configure the LLVM build using the `-DLLVM_ENABLE_RTTI`
flag. It would be nice is this type of stuff was documented in the tutorial, but
maybe I should just make my own tutorial which is the fixed version.